Contain mask component within HiddenMod
