+++ osu.Game.Rulesets.Sentakki/Mods/SentakkiModHidden.cs
@@ -1,4 +1,13 @@
+using System;
+using osu.Framework.Allocation;
 using osu.Framework.Graphics;
+using osu.Framework.Graphics.Batches;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.OpenGL.Vertices;
+using osu.Framework.Graphics.Primitives;
+using osu.Framework.Graphics.Shaders;
+using osu.Framework.Graphics.Textures;
+using osu.Game.Graphics.OpenGL.Vertices;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Rulesets.Objects.Drawables;
 using osu.Game.Rulesets.Scoring;
@@ -6,6 +15,7 @@ using osu.Game.Rulesets.Sentakki.Objects;
 using osu.Game.Rulesets.Sentakki.Objects.Drawables;
 using osu.Game.Rulesets.Sentakki.UI;
 using osu.Game.Rulesets.UI;
+using osuTK;
 
 namespace osu.Game.Rulesets.Sentakki.Mods
 {
@@ -15,7 +25,7 @@ namespace osu.Game.Rulesets.Sentakki.Mods
 
         public override double ScoreMultiplier => 1.06;
 
-        public virtual void ApplyToDrawableRuleset(DrawableRuleset<SentakkiHitObject> drawableRuleset)
+        public void ApplyToDrawableRuleset(DrawableRuleset<SentakkiHitObject> drawableRuleset)
         {
             SentakkiPlayfield sentakkiPlayfield = (SentakkiPlayfield)drawableRuleset.Playfield;
             LanedPlayfield lanedPlayfield = sentakkiPlayfield.LanedPlayfield;
@@ -24,7 +34,7 @@ namespace osu.Game.Rulesets.Sentakki.Mods
             var lanedNoteProxyContainer = lanedHitObjectArea.Child;
 
             lanedHitObjectArea.Remove(lanedNoteProxyContainer);
-            lanedHitObjectArea.Add(new CircularPlayfieldCoveringWrapper(lanedNoteProxyContainer, VisibilityReductionMode.Hidden)
+            lanedHitObjectArea.Add(new PlayfieldMaskingContainer(lanedNoteProxyContainer)
             {
                 CoverageRadius = 0.4f
             });
@@ -59,6 +69,144 @@ namespace osu.Game.Rulesets.Sentakki.Mods
                     using (sb.BeginAbsoluteSequence(sb.HitObject.StartTime - preemptTime))
                         sb.Slidepath.FadeOut(fadeOutTime);
                     break;
+            }
+        }
+
+        private class PlayfieldMaskingContainer : CompositeDrawable
+        {
+            private readonly PlayfieldMask cover;
+
+            public PlayfieldMaskingContainer(Drawable content)
+            {
+                RelativeSizeAxes = Axes.Both;
+
+                InternalChild = new BufferedContainer
+                {
+                    RelativeSizeAxes = Axes.Both,
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    Children = new Drawable[]
+                    {
+                        new Container
+                        {
+                            RelativeSizeAxes = Axes.Both,
+                            Anchor = Anchor.Centre,
+                            Origin = Anchor.Centre,
+                            Child = content
+                        },
+                        cover = new PlayfieldMask
+                        {
+                            Anchor = Anchor.Centre,
+                            Origin = Anchor.Centre,
+                            RelativeSizeAxes = Axes.Both,
+                            Blending = new BlendingParameters
+                            {
+                                // Don't change the destination colour.
+                                RGBEquation = BlendingEquation.Add,
+                                Source = BlendingType.Zero,
+                                Destination = BlendingType.One,
+                                // Subtract the cover's alpha from the destination (points with alpha 1 should make the destination completely transparent).
+                                AlphaEquation = BlendingEquation.Add,
+                                SourceAlpha = BlendingType.Zero,
+                                DestinationAlpha = BlendingType.OneMinusSrcAlpha
+                            },
+                        }
+                    }
+                };
+            }
+
+            /// <summary>
+            /// The relative area that should be completely covered if it is FadingIn, or the visible area if it is Hidden.
+            /// Doesn't include fade
+            /// </summary>
+            public float CoverageRadius
+            {
+                set => cover.ApertureSize = new Vector2(0, SentakkiPlayfield.RINGSIZE / 2 * value);
+            }
+
+            private class PlayfieldMask : Drawable
+            {
+                private IShader shader;
+
+                protected override DrawNode CreateDrawNode() => new PlayfieldMaskDrawNode(this);
+
+                [BackgroundDependencyLoader]
+                private void load(ShaderManager shaderManager)
+                {
+                    RelativeSizeAxes = Axes.Both;
+                    Anchor = Origin = Anchor.Centre;
+                    shader = shaderManager.Load("PositionAndColour", "PlayfieldMask");
+                }
+
+                private Vector2 apertureSize;
+
+                public Vector2 ApertureSize
+                {
+                    get => apertureSize;
+                    set
+                    {
+                        if (apertureSize == value) return;
+
+                        apertureSize = value;
+                        Invalidate(Invalidation.DrawNode);
+                    }
+                }
+
+                public Vector2 AperturePosition => ToParentSpace(OriginPosition);
+
+                private class PlayfieldMaskDrawNode : DrawNode
+                {
+                    protected new PlayfieldMask Source => (PlayfieldMask)base.Source;
+
+                    private IShader shader;
+                    private Quad screenSpaceDrawQuad;
+
+                    private Vector2 aperturePosition;
+                    private Vector2 apertureSize;
+
+                    private readonly VertexBatch<PositionAndColourVertex> quadBatch = new QuadBatch<PositionAndColourVertex>(1, 1);
+                    private readonly Action<TexturedVertex2D> addAction;
+
+                    public PlayfieldMaskDrawNode(PlayfieldMask source)
+                        : base(source)
+                    {
+                        addAction = v => quadBatch.Add(new PositionAndColourVertex
+                        {
+                            Position = v.Position,
+                            Colour = v.Colour
+                        });
+                    }
+
+                    public override void ApplyState()
+                    {
+                        base.ApplyState();
+
+                        shader = Source.shader;
+                        screenSpaceDrawQuad = Source.ScreenSpaceDrawQuad;
+                        aperturePosition = Vector2Extensions.Transform(Source.AperturePosition, DrawInfo.Matrix);
+                        apertureSize = Source.ApertureSize * DrawInfo.Matrix.ExtractScale().Xy;
+                    }
+
+                    public override void Draw(Action<TexturedVertex2D> vertexAction)
+                    {
+                        base.Draw(vertexAction);
+
+                        shader.Bind();
+
+                        shader.GetUniform<Vector2>("aperturePos").UpdateValue(ref aperturePosition);
+                        shader.GetUniform<Vector2>("apertureSize").UpdateValue(ref apertureSize);
+
+                        DrawQuad(Texture.WhitePixel, screenSpaceDrawQuad, DrawColourInfo.Colour, vertexAction: addAction);
+
+                        shader.Unbind();
+                    }
+
+                    protected override void Dispose(bool isDisposing)
+                    {
+                        base.Dispose(isDisposing);
+                        quadBatch?.Dispose();
+                    }
+                }
             }
         }
     }
